feature: refactor project to separate logics into specific files
